Alterei o sítio de alguns métodos
Alterei o sítio de métodos relacionados com os checkpoints para estar em concordância com os scripts PlayerController e PlayerMovement.

Alterei o script HoleTrigger para usar um PlayerMovement em vez de PlayerController
