--- jogo-base/Library/ScriptAssemblies/BuiltinAssemblies.stamp
@@ -1,2 +1,0 @@
-0000.5a6f802e.0000
-0000.5a6f94e2.0000
